third_party/WebKit/LayoutTests/fast/forms/pattern-attribute-invalid-regex.html

<!DOCTYPE html>

<script src="../../fast/js/resources/js-test-pre.js"></script>

<script>
description('Check if the pattern constraint is ignored when the regex is invalid.');

var input = document.createElement('input');
input.pattern = '[0-9';  // missing a ]
input.value = '11';

// pattern should have been ignored.
shouldBe('input.validity.patternMismatch', 'false');
</script>

<script src="../../fast/js/resources/js-test-post.js"></script>
