third_party/WebKit/LayoutTests/fast/forms/pattern-attribute-invalid-regex-expected.txt

Check if the pattern constraint is ignored when the regex is invalid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.validity.patternMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
